Change the C++ code to utilize the GPSDO for timing purposes. There is much to add regarding the GPSDO, this change merely keeps the code functional.
Why was this change needed?
During installation of the GPSDO, a physical switch inside the N200 needed to be changed that switches the N200 from using its own internal clock to using the GPSDO clock. The 'ref' variable in the C++ code was defaulted to 'internal', as before now there was no GPSDO to take over timing, but because of that physical hardware switch, the internal clock of the N200 no longer works. This caused the program to be nonfunctional.

This change fixes that, and data is successfully collected using the GPSDO to time the ticks inside the N200.